feat: Upload published packages to specified cache
If a cache is specified with `flox publish --cache`, we now copy the
published artifacts with `nix copy --to <cache>`. The packages are
signed using the key as specified by the new `--key-file` flag, which is
required if `--cache` is provided.

The binary cache logic is encapsulated by a new BinaryCache trait, along
with a mock implementation for testing.