feat: allow overriding Nix store
This allows a user to override the Nix store via the _FLOX_NIX_STORE_URL
variable. This is done in both `pkgdb realise` and on the Rust side. The
Rust side happens in BuildEnvNix::base_command by reading the
environment variable, which we typically don't allow in the SDK. I'm
doing it now to unblock another developer and this code will be
superseded by another ticket in the very near future.