feat: realise catalog packages from nixpkgs
test: add unit tests for building packages from the nixpkgs catalog

chore: generate lockfiles for unfree and broken packages

* unfree -> `hello-unfree`
Ensure that we can build, or (attempt to build) a package from the catalog,
that is marked as **unfree**.
By default, unfree packages are included in resolution responses,
unless explicitly disabled.
Nixpkgs provides an _evaltime_ check for this metadata attribute,
causing evaluation failures unless configured otherwise.
Since we have our own control mechanism and generally want to skip evaluations
if possible, we rely on [[BuildEnvNix::realise_nixpkgs]]
to successfully evaluate the package and build it.

* broken -> `tabula`
Ensure that we can build, or (attempt to build) a package from the catalog,
that is marked as **broken**.
marked as broken may build successfully, but are not guaranteed to work.
By default, the packages are not included in resolution responses,
unless explicitly enabled.
Nixpkgs provides an _evaltime_ check for this metadata attribute,
causing evaluation failures unless configured otherwise,.
Since we have our own control mechanism and generally want to skip evaluations
if possible, we rely on [[BuildEnvNix::realise_nixpkgs]]
to (at least) successfully evaluate the package, and attempt to build it.