feat: realise packages from flakes
The package is realised by checking if the store paths are valid,
and otherwise building the package to create valid store paths.
Packages are built by optimistically joining the flake url and attr path,
which has been previously evaluated successfully during locking,
and building the package with essentially `nix build <flake-url>#<attr-path>^*`.
We set `--option pure-eval true` to avoid improve reproducibility,
and allow the use of the eval-cache to avoid costly re-evaluations.

test: add unit tests for building packages from flakes