refactor: replace `pkgdb realise` with nix call implementation
fix(buildenv): err on incompatible lockfile

Check if the lockfile is compatible with the current system.
Explicitly setting the `options.systems` field in the manifest,
has the semantics of restricting the environments to the specified systems.
Restricting systems can help the resolution process and avoid confusion,
when using the environment on unsupported systems.
Without this check the lockfile would succeed to build on any system,
but (in the general case) contain no packages,
because the lockfile won't contain locks of packages for the current system.