feat: add `lock-manifest` command
Adds a new hidden and strictly internal command `flox lock-manifest`,
that allows locking of a manifest, separately from an environment.
With `--lockfile` allows to specify a base lockfile.
This command _only_ locks and never builds the resulting lockfile,
which makes it useful for generating test data,
in particular for the use with the environment building subsystem.