feat(build): set `HOME` to a writable directory in the build sandbox
In the Nix sandbox $HOME is usually set to `/homeless-shelter`,
does not exist, and cannot be written to.
In turn, any tool attempting to write to $HOME will experience errors to do so.
We set $HOME to another writable location in the sandbox,
to ensure such errors do not occur.

See also: <https://github.com/flox/flox/issues/2092>