feat(build): detect empty $out/bin and executables outside
Only executables in `$out/bin` will be available on the PATH and wrapped correctly.
When a build creates files in $out but does not place any files in $out/bin,
we want to warn that Flox needs binaries in bin/, so users can correct the build command accordingly.

Now, when '$out/bin' does not contain any executable files, the build will print the following warning:

```
⚠️  WARNING: No executables found in '$out/bin'.
Only executables in '$out/bin' will be available on the PATH.
If your build produces executables, make sure they are copied to '$out/bin'.
  - copy a single file with 'mkdir -p $out/bin && cp file $out/bin'
  - copy a bin directory with 'mkdir $out && cp -r bin $out'
  - copy multiple files with 'mkdir -p $out/bin && cp bin/* $out/bin'
  - copy files from an Autotools project with 'make install PREFIX=$out'
```

If in addition there are executable files anywhere else in `$out`,
the build will provide a hint to inform which files to potentially move:

```
HINT: The following executables were found outside of '$out/bin':
  - not-bin/hello
```