test: reimplement `buildenv.bats` tests to rust unit tests
`buildenv.bats` depends on `pkgdb realise`
to realise the input environments built with `buildenv.nix`.
To drop the dependency on `pkgdb` and bring the tests closer to the provider
that actually drives the implementation,
as well as avoiding additional process spawning,
this commit aims to convert the existing bats tests 1-to-1 to rust.

chore(test_data): generate lockfiles used by the buildenv tests,
where resolution is required