feat: add "log" output
Adding a new "log" output for recording the timestamped, color-highlighted
output produced by the build. In the case of pure builds this is simply a
new output, but in the case of impure builds we need to make this a separate
derivation rather than having the logfile be an ever-changing input to the
nix build which creates the "out" result. This output will be consumed by
`flox publish` when uploading packages and their associated logs to the
binary cache.