feat: echo lines in the build script as they are invoked
It is almost always the case that users will want to see the lines of
their [manifest] build script echo'd as it runs. Keep in mind that this
is not the build itself, but rather the minimal set of lines required
to kick off the build, and doing it this way also ensures that those
lines are highlighted as they are written to stderr.