refactor: define "container builder" as a provider trait
The general idea is that a container builder implements the means to create a `ContainerSource`.
The `ContainerSource` is a script that writes a container tarball to stdout.
The source can be tapped and written to a `sink : &mut impl Write`.

This architecture follows from the way that nixpkgs' `dockerTools` are designed.
`nixpkgs#dockerTools.streamLayeredImage` produces a script that composes store paths
as layers of a container image, and writes the resulting image to stdout.
The main intention is to avoid "building" the image in nix terms,
writing it to the nix store from where it would immediately copied into a container engine,
doubling the memory consumed.