docs(edit): Clarify docs for --sync and --reset
The error and the docs confuse me each time I run into them because I'm
not immediately sure about which direction the sync or reset are
operating and I don't have to interact with the concept of generations
very often.

Attempt to clarify this by using the terms "commit" and "discard" from
the perspective of local only, which is where the changes have
originated from and what the user may lose. This doesn't appear to
conflict with the original design in:

- https://github.com/flox/flox/issues/680

I've omitted the "after modifying manifest.toml" part because by
definition the manifest has already been modified when you've reached
this error and we don't have any notion of conflict resolution until we
add diffs in:

- https://github.com/flox/flox/issues/1634