feat(containerize): Add support for macOS
By using the proxy container approach from:

- https://www.notion.so/floxdev/012-Containerize-cross-platform-14a744c8e3168090b508dafd1fc9f37d0

This has some rough edges which we'll polish later..

The proxy container has to fetch/build store paths on every run. We can
reduce this time by caching/persisting the proxy container in 2469. It
may also be desirable to separate the build of the Flox flake from the
build of the environment for better feedback, but it's not worth doing
that until we refactor for caching.

Testing has only been done manually so far because we don't have a
functioning Docker or Podman in CI. This will be covered in 2466.

Output from `nix run` interleaves the "Writing container" spinner but it
does give some useful feedback about why it might be slow and is
consistent with the existing behaviour when writing layers.

The version of the Flox flake is pinned to a specific commit if the
semver resolves to v1.3.6 without a commit ref, as it does with `just
build-cli`, because it doesn't contain the STDOUT fix. It can be removed
once we release v1.3.7.

I've tested that it also works for remote environments.

Creating a `ContainerSource` for macOS doesn't require a spinner because
it doesn't execute anything until `stream_container()`.