fix: only run profile for repeat Bash activations
For in-place and command activations, we don't error when an environment
is already active, but we instead perform a normal activation. This
causes problems for activations in RC files.

If `eval "$(flox activate -d default)"` is in an RC file and then
another environment is activated with `flox activate -d project`, the
default environment is put before the project environment in PATH
because the activation of the project environment sources RC files as
part of activation.

To avoid this problem, only run profile steps for repeated in-place
activations. Don't set PATH or MANPATH, start services, or set variables
from `hook.on-activate`.

We can't skip repeated activations entirely. If we did that, in the
example above, any aliases set for the default environment would not be
set in the user's shell after running `flox activate -d project`.