build: consume internal dependencies by dynamic path in devshells
Out codebase has grown multiple subsystems that are depended on by flox
or depend on other subsystems themselves.
E.g. `flox -> pkgdb -> flox-activation-scripts -> flox-activations` for one of the worse ones.
On the one hand it is important to be able to build final artefacts
of each individual subsystem with nix to allow for better build caching if nothing else.
Nix also makes defining these dependencies relatively easy through e.g. string contexts.

However, since our devshell is derived from _the inputs of the toplevel binaries_,
the naive approach results in
1. all dependencies being built upon entering the devshell
2. the devshell being in accessible while any dependency fails to build
3. the devshell being stale as soon as any dependency changes -- requiring us to reenter the devshell

Until not too long ago, most of the now standalone subsystems were consolidated in pkgdb
(i.e. buildenv and associated functionality).
In order to avoid the 3 pitfalls above, for development, we would set `PKGDB_BIN` to `pkgdb`,
instead of `${flox-pkgdb}/bin/pkgdb`, in turn removing the dependency on pkgdb form the build closure,
as well as the dev shell closure.
`Pkgdb` now needed to be built separately from the `flox` binary,
which we ensured would happen through via the central Justfile.
As we added internal (nix built) dependencies to `pkgdb`, we amended the pkgdb Makefile
to build these dependencies as requisites of the `pkgdb` binary
and provide their path to the C++ preprocessor, when they were unspecified in dev mode.

While `flox` only depended on binaries this worked out relatively well,
except occasionally we would still find internal dependencies in the closure of the devshell.
That was because we

1. forgot to add "dev mode" overrides,
2. depend on a package by both `flox` and `pkgdb`,
3. required absolute paths which broke our heuristic of adding binaries to PATH,
4. etc.

As we are phasing out pkgdb as a binary, and move more functionality into independent packages
orchestrated by the `flox` binary, we are losing some of the dynamic setup
we had implemented via the `pkgdb` Makefile.
For example, most recently we have experienced issues wrt `flox-activation-scripts`,
which made its way into the devshell (alongside, flox-buildenv and transitively `pkgdb` ...).

This commit is trying to consolidate the dependency management for `flox`
and the remains of pkgdb, by applying a similar strategy at an earlier stage.

In short

1. packages **only** provide build environment variables that reference dependencies
   iff the dependency is provided, i.e. not overridden to `null`
2. within the `floxDevelopmentPackages` scope,
   all internal dependencies of internal packages are overridden to `null.
3. define all environment variables that point to internal dependencies
   within the `shellHook` of the devshell and point them to a local directory.
   `cargo` built dependencies become: `FLOX_ACTIVATIONS_BIN="''${REPO_ROOT}/cli/target/debug/flox-activations"`
   `nix` built dependencies become `FLOX_BUILDENV="''${REPO_ROOT}/build/flox-buildenv"`
4. define Justfile targets that build dependencies with either `cargo` or `nix`
   thereby populating the directories previously defined in the environment.

Compared to the previous Makefile implementation this avoids requiring to rebuild binaries
as the dependencies change, because everything will depend on stable paths that are updated in place.

squash! build: consume internal dependencies by dynamic path in devshells

ci: include `bats` in devshell

squash! build: consume internal dependencies by dynamic path in devshells

remove `flox-activation-scripts` from closure of `flox-cli-tests` within devshell