feat: recommend .profile for default env
On macOS every shell is started as a login shell, which is different
from many Linux distributions. This means that the recommendation to put
the default environment activation in `.bashrc` will not activate the
default environment since Bash only loads `.profile` for login shells,
not `.bashrc`. To address this we put the activation line in both
scripts.

No tests were added because we already have tests for verifying that
individual RC files were modified, but the new functionality and
messaging was verified interactively.