fix(activate): nested activation for zsh
Rather than generating a startup script as for Bash in f623518, extend
the approach added to .zshrc in 52df88e that saves and restores
variables used in activation. Extend it to other zsh startup files, and
also add _FLOX_RESTORE_PATH, _FLOX_RESTORE_MANPATH, and
_FLOX_ACTIVATION_PROFILE_ONLY to the save and restore logic.

The current approach could be improved by:
improved by:
- looping over variable names
- sourcing a common file and function from all 4 scripts

Respect _FLOX_ACTIVATION_PROFILE_ONLY in the zsh startup script, similar
to what was added for Bash in fa66016.