feat(containerize): Proxy more specific commit
Use the `flox/flox` flake at a more specific commit SHA, if available.

In practice this might not actually make any difference when testing a
locally built binary (as will be used in CI) because we don't append the
version:

   % which flox
   /Users/dcarley/projects/flox/flox/cli/target/debug/flox

   % flox --version
   1.3.6

Unlike:

    % nix run github:flox/flox -- --version
    1.3.7-g6433689

So it will likely fallback to the base semver until we fix the linked
ticket.