build: add a wrapper for use outside of the dev shell
When built for use in development, absolute paths to nix store paths in nix built scripts
are replaced for references to environment variables that are provided in the devshell.
If we attempt to execute `flox` from outside of the devshell, these variables are missing
and may cause issues at runtime.

This Commit adds a wrapper script `<repo>/build/wrapper` which executes any given command
with the environment dev variables set, allowing the use of flox outside of the devshell.

```shell
$ flox develop -c true # build and immediately exit the devshell
$ ./build/wrapper flox <flox args...>
```