chore: Remove unnecessary disambiguate code
Previously, we expected that multiple catalogs would be able to provide
a package with the same name with install-time resolution deciding which
catalog to use. We foresaw a need in that case to disambiguate by
prepending the catalog name, only when necessary. In reality, catalogs
contribute a prefix to a single global namespace, meaning there is never
a case where we might have two packages with the same name and
therefore the disambiguation code is unnecessary.