fix: Properly account for highlighting while aligning table
When determining the width for the 'name' column, we were counting
non-printable color codes. That led to a misaligned table in the case
where some package names didn't include the search term and thus were
unhighlighted. We now determine the expected column width based on the
unformatted name, and then account for the extra color code characters
only when printing a name that includes them.