test: enable macos containerize tests
This starts the Podman VM at the beginning of the `containerize.bats`
file setup, and reuses it for all of the tests in this file. This was
tricky because Podman uses a deeply nested directory structure under XDG
directories, which in some cases created socket paths that were much
longer than 108 characters. One of these files is placed under a
hardcoded path in the home directory, which means that all of the tests
must share a single home directory in order to share the VM.

Another interesting detail is that certain directories are not available
in the VM unless you explicitly mount them in. The Docker VM mounts these
automatically, but we need to mount them manually.

Running these tests is disabled on the Github macOS runner because you
can't double nest VMS. It's also disabled for `x86_64-darwin` because
those tests are run under Rosetta on an `aarch64-darwin` machine, which
causes problems.

Finally, debug logging for the startup of the Podman machine is disabled
because the test suite hangs after completion for whatever reason.