fix: ensure `mkContainer.nix` is in the closure of flox
The path to the `mkContainer.nix` script was compiled into the flox binary
starting with https://github.com/flox/flox/pull/2473.
However, upon inspection of the closure,
the corresponding store path does not seem to be part of the closure itself.

Adding `flox-mk-container` to the package closure explicitly via `propagatedBuildInputs`.