fix: drop redundant "Producing..." spinners
This drops steps from spinners that don't add useful information.

For checking if a package has already been downloaded, don't say the
package is being produced or ensured to be present. Although we may skip
downloading if the store path is present, in that case the spinner will
only flash briefly, and if the user reads it they will understand what
happened - the download was fast since it was cached.

Before:
```
⠐ Installing 1 package(s) to environment 'main'...
  ↳ ⠄ Building environment
    ↳ ⠄ Producing package 'nodejs'
      ↳ ⠄ Ensuring packages are present or downloaded
```

After:
```
⠤ Installing 1 package(s) to environment 'main'...
  ↳ ⠒ Building environment
    ↳ ⠂ Downloading 'nodejs'
```

Drop the "Producing..." line when building from source as well:

Before:
```
⠒ Installing 1 package(s) to environment 'main'...
  ↳ ⠄ Building environment
    ↳ ⠄ Producing package 'vscode'
      ↳ ⠄ Building 'vscode' from source
```

After:
```
⠁ Installing 1 package(s) to environment 'main'...
  ↳ ⠐ Building environment
    ↳ ⠦ Building 'vscode' from source
```

For store paths, realising and ensuring the package is present or
downloaded mean exactly the same thing, so drop the line about
"Ensuring..."

Before:
```
⠤ Installing 1 package(s) to environment 'main'...
  ↳ ⠤ Building environment
    ↳ ⠴ Realising store path for 'nodejs'
      ↳ ⠴ Ensuring packages are present or downloaded
```

After:
```
⠁ Installing 1 package(s) to environment 'main'...
  ↳ ⠐ Building environment
    ↳ ⠦ Realising store path for 'nodejs'
```

For flake packages, drop the `Checking if store paths exist locally`
spinner since it will be nearly instantaneous and we already have a
parent `Realising flake package '{}'` spinner.