fix: share home and xdg dirs for file
`podman` starts a few different processes in order to run both the VM
and support port mapping between the host and the container. These
include `gvproxy` and `vfkit`, where the former manages DNS/ports, and
the latter runs the VM.

These processes store files and sockets in both XDG directories _and_ the
home directory. This means that in order to support a single VM for the
entire file, we need to share the home directory and XDG directories.
The sockets created under these directories can have paths longer than
108 characters, so temporary directories are created directly under
`/tmp` to back the home directory and the XDG dirs. On top of that, the
`nix develop` command sets `TMP`, `TMPDIR`, etc to a temporary directory
created for the devshell. That can also cause socket paths to be too
long, so we also set `TMPDIR` back to a path directly under `/tmp`.