fix: wrap binaries in libexec, fix builder warnings
The previous approach for surveying the manifest build output was
showing false positives, and it's long been the case that we weren't
properly handling executables in libexec either.

This patch addresses both of these problems by generally being more
precise about the executables to be wrapped in the following ways:

- warns about
  - non-files in {bin,sbin} (note we do not warn for libexec)
  - non-executables in {bin,sbin} (note we do not warn for libexec)
  - no executable files found in bin
  - executable files in directories other than {bin,sbin,libexec},
    including subdirectories of {bin,sbin,libexec}
- no longer attempts to wrap subdirectories of {bin,sbin,libexec}
- fixes bug in printing of "stray" executables found in package

This patch also renames and consolidates the testing of the "sanity"
of the out package into a single `build_verify_sane_out()` function
to be called in both sandbox modes, and replaces the use of `fd`
with `find` to cut down on the runtime closure.