feat: add provider for upgrade checks
Add an access layer for an `UpgradeInformation` file.
An `UpgradeInformationGuard<Unlocked>` is used to access a possibly absent file.
The guard can attempt to acquire a file lock to gain exclusive access to the `UpgradeInformation` on disk.
`UpgradeInformationGuard::try_lock` will _not_ block, but instead return early
without changing state.