feat: add hidden `check-for-upgrades` command
A simple subcommand that opens a provided `UninitializedEnvironment` and attempts a `dry_upgrade`.
The result is written to an `UpgradeInformation` file.
Concurrent writes to the file are prevented by first acquiring a lock.
If there already is an UpgradeInformation file, which has been checked less than `<check-timeout>` seconds (default 1 day) ago,
or any other process currently owns a lock,
the process exits early without attempting another upgrade, or touching the file.