feat: flox-path-helper for grooming PATH & MANPATH
Reimplemented PATH and MANPATH rewriting logic by way of a flox-path-helper,
patterned in a way after the Mac path-helper. The output of this script can
be sourced at any time to set the PATH and MANPATH variables according to the
current value of the FLOX_ENV_DIRS variable.

This script allowed us to remove a bunch of logic associated with the setting
of PATH and MANPATH, as well as retiring the previous set-run-variables.bash
script in its entirety.

Also tangentially addressed by way of this patch:

- chore: update tracer invocations in activate script to identify invocation path
- chore: be consistent in generate-* scripts with which variables get passed
  down by different shells
- chore: remove FLOX_ENV_DIRS FLOX_ENV_LIB_DIRS logic from Rust CLI