feat: add FLOX_ACTIVATE_TRACE debugging feature
With the myriad of activation scripts I was finding it challenging to track
environment variable changes happening in the different files, so I added
invocations of a $FLOX_ACTIVATE_TRACE script at the beginning and end of
each file in order to trace their various entry/exit points. While this PR
includes a default trace script, it also provides the ability for people to
provide their own, making it a convenient hook for inspecting the changes
that can happen to arbitrary variables.

The full instructions for using this variable have been added to the
CONTRIBUTING.md document.