feat: add helper function for spawning an async upgrade check
Spawn a new `flox check-for-upgrades` process in the background,
and redirect its logs to a log file.

The process will live on after the parent process exits.
When multiple processes are spawned, e.g. due to multiple successive activations,
one (usually the first) process will grab a lock on the upgrade information file,
and the others will exit early.