feat(activate): notify users applicable upgrade information was found
Before activation, read the result of a previous `check-for-upgrade`.
If an UpgradeInfo file exists, corresponds to the activated lockfile,
and suggests that upgrades are available,
compile a report of what packages could be upgraded.

feat: spawn async check for upgrades before executing the activation

feat: add config option to disable the display of upgrade checks