fix: do not expect flox-cli-tests PATH in activated envs
The jq_edit() function implicitly relied on finding `jq` in the PATH but when used
within an activated environment not containing the `jq` package could not find it.
This worked previously because our approach to "restoring" the PATH within a nested
activation had the effect of cancelling changes to the PATH that can happen in user
dotfiles, but now that we have fixed that we now need to be explicit about making
available _all_ programs required by a test, and there are two ways to do this:

1. add all required programs to the test "project" environments
   - this requires updating mock data in addition to adding a "flox install"
     for adding the missing packages
2. pass the explicit path to the required utility from the flox-cli-tests
   environment (which we know exists) down to the test, e.g. by defining a
   new `$_jq` variable for passing the full path to `jq`

I've elected to do the second of these for tests that needed access to such paths,
but I'm happy if someone prefers to revise this and use the first approach.

Again, this problem did not surface previously because our previous approach of
restoring the previous PATH had the effect of pulling the test runner's PATH into
the test. This patch updates jq_edit() to use the $_jq env variable if defined to
identify the full path to jq, and then exercises it from the [1] test where it
won't otherwise have access to jq in the PATH. It applies a similar solution for
locating the path to `man` and `expect` in other tests.