fix: always activate using the `--env` flag
Updating all invocations of the `activate` script to use the explicit
`--env` flag for specifying the environment to be activated. We do
this because we don't want to rely on FLOX_ENV from the environment
a) when we explicitly know what it should be and b) when we are performing
an "inner" activation of a nested environment and it may already have
been defined (and exported) by an "outer" activation.