fix: reinstate activate:inplace-reactivate tests
This reverts commit 6128bbb5ec723b58363d90c0a74032dae40a2a46, which itself
removed tests that were created to confirm the correct PATH after repeated
in-place activations of an already-active environment.

In that commit Matthew also explained that the tests weren't actually testing
what we thought it was because bats was expanding a set of quotes on the way
to invoking the test. This patch addresses that problem by avoiding quotes
altogether, and instead sourcing the output of `flox activate` by way of
a file rather than an eval'd string.

Matthew's commit also commented that "we're changing behavior" with regard
to the setting of PATHs for nested activations, but PR 2552 restores that
behavior so it makes sense to bring back these tests as well.