build: use meson to build `nix-plugins`
To me the explicitness, and use of _types_ made the definition of the build using meson
more concise and clear -- i take that this is personal judgement so far.
In addition to explicitness, meson also provides integration with editors when needed,
a structure for defining and running tests, a pkg-config interface, etc,
which are usually (and were previously) hand rolled.
In absolute numbers, we exchanged ~700 lines of make, for ~100 lines of meson.
I don't doubt that a minimal makefile would be even less,
bar most of the benefits of the meson variant.