feat: source profile hooks for all active environments
When nesting environment activations the relevant `profile` hooks for all
activated environments should be sourced when starting a new subshell.
Prior to this patch only the profile hooks for the current / most recent
environment was being sourced, and this patch adds logic to iterate over
the contents of the FLOX_ENV_DIRS variable when looking for profile hook
scripts to be sourced.

This patch also adds tests to confirm that profile hooks are sourced in the
right order and for all active environments in a nested activation. Note that
these test use the recently-added _FLOX_SHELL_FORCE variable.

Co-authored-by: Matthew Kenigsberg <matthew@floxdev.com>